+++ test/lib/bamboo/plug/sent_email_viewer_plug_test.exs
@@ -1,4 +1,4 @@
-defmodule Bamboo.EmailPreviewTest do
+defmodule Bamboo.SentEmailViewerPlugTest do
   use ExUnit.Case
   use Plug.Test
   import Bamboo.Factory
@@ -10,8 +10,8 @@ defmodule Bamboo.EmailPreviewTest do
     plug :match
     plug :dispatch
 
-    forward "/sent_emails/foo", to: Bamboo.EmailPreviewPlug
-    forward "/", to: Bamboo.EmailPreviewPlug
+    forward "/sent_emails/foo", to: Bamboo.SentEmailViewerPlug
+    forward "/", to: Bamboo.SentEmailViewerPlug
   end
 
   setup do
@@ -19,16 +19,17 @@ defmodule Bamboo.EmailPreviewTest do
     :ok
   end
 
-  test "shows list of all sent emails, and previews the newest email" do
+  test "shows list of all sent emails, and the body of the newest email" do
     emails = normalize_and_push_pair(:email)
     conn = conn(:get, "/sent_emails/foo")
 
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
-    assert selected_sidebar_email_text(conn) =~ newest_email.subject
-    assert showing_in_preview_pane?(conn, newest_email)
-    refute showing_in_preview_pane?(conn, oldest_email)
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
+    assert selected_sidebar_email_text(conn) =~ newest_email().subject
+    assert showing_in_detail_pane?(conn, newest_email())
+    refute showing_in_detail_pane?(conn, oldest_email())
     for email <- emails do
       assert Floki.raw_html(sidebar(conn)) =~ ~s(href="/sent_emails/foo/#{SentEmail.get_id(email)}")
       assert Floki.text(sidebar(conn)) =~ email.subject
@@ -46,10 +47,52 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert conn.resp_body =~ Bamboo.Email.get_address(email.from)
     for email_address <- Bamboo.Email.all_recipients(email) do
       assert conn.resp_body =~ Bamboo.Email.get_address(email_address)
     end
+  end
+
+  test "prints single header in detail pane" do
+    email = normalize_and_push(:email, headers: %{"Reply-To" => "reply-to@example.com"})
+    conn = conn(:get, "/sent_emails/foo")
+
+    conn = AppRouter.call(conn, nil)
+
+    assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
+    assert conn.resp_body =~ Bamboo.Email.get_address(email.from)
+    assert conn.resp_body =~ "Reply-To"
+    assert conn.resp_body =~ "reply-to@example.com"
+  end
+
+  test "prints multiple headers in detail pane" do
+    email = normalize_and_push(:email, headers: %{"Reply-To" => ["reply-to1@example.com", "reply-to2@example.com"], "Foobar" => "foobar-header"})
+    conn = conn(:get, "/sent_emails/foo")
+
+    conn = AppRouter.call(conn, nil)
+
+    assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
+    assert conn.resp_body =~ Bamboo.Email.get_address(email.from)
+    assert conn.resp_body =~ "Reply-To"
+    assert conn.resp_body =~ "reply-to1@example.com"
+    assert conn.resp_body =~ "reply-to2@example.com"
+    assert conn.resp_body =~ "Foobar"
+    assert conn.resp_body =~ "foobar-header"
+  end
+
+  test "falls back to inspect when printing header value that is not a string or list" do
+    normalize_and_push(:email, headers: %{"SomeHeader" => %{"Some" => "Header"}})
+    conn = conn(:get, "/sent_emails/foo")
+
+    conn = AppRouter.call(conn, nil)
+
+    assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
+    assert conn.resp_body =~ "SomeHeader"
+    assert conn.resp_body =~ "%{\"Some\" => \"Header\"}"
   end
 
   defp selected_sidebar_email_text(conn) do
@@ -63,6 +106,7 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert Floki.raw_html(sidebar(conn)) =~ ~s(href="/#{SentEmail.get_id(email)}")
     assert Floki.text(sidebar(conn)) =~ email.subject
   end
@@ -73,6 +117,7 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert conn.resp_body =~ "No emails sent"
   end
 
@@ -85,8 +130,9 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
-    assert showing_in_preview_pane?(conn, SentEmail.get(selected_email_id))
-    refute showing_in_preview_pane?(conn, SentEmail.get(unselected_email_id))
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
+    assert showing_in_detail_pane?(conn, SentEmail.get(selected_email_id))
+    refute showing_in_detail_pane?(conn, SentEmail.get(unselected_email_id))
   end
 
   test "shows an email's html by id" do
@@ -97,6 +143,7 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert conn.resp_body =~ SentEmail.get(selected_email_id).html_body
   end
 
@@ -108,6 +155,7 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 200
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert conn.resp_body == ""
   end
 
@@ -117,6 +165,7 @@ defmodule Bamboo.EmailPreviewTest do
     conn = AppRouter.call(conn, nil)
 
     assert conn.status == 404
+    assert {"content-type", "text/html; charset=utf-8"} in conn.resp_headers
     assert conn.resp_body =~ "Email not found"
   end
 
@@ -128,12 +177,12 @@ defmodule Bamboo.EmailPreviewTest do
     SentEmail.all |> List.last
   end
 
-  defp showing_in_preview_pane?(conn, email) do
-    Floki.text(preview_pane(conn)) =~ email.subject
+  defp showing_in_detail_pane?(conn, email) do
+    Floki.text(detail_pane(conn)) =~ email.subject
   end
 
-  defp preview_pane(conn) do
-    conn.resp_body |> Floki.find(".email-preview-pane")
+  defp detail_pane(conn) do
+    conn.resp_body |> Floki.find(".email-detail-pane")
   end
 
   defp sidebar(conn) do